+++ lib/elastic_client.dart
@@ -1,3 +1,4 @@
 export 'src/elastic_client_impl.dart';
+export 'src/http_transport.dart';
 export 'src/query.dart';
 export 'src/transport.dart';