+++ lib/src/http_transport.dart
@@ -1,20 +1,30 @@
-import 'dart:async';
+import 'dart:convert' as convert;
 
-import 'package:http_client/http_client.dart' as http;
+import 'package:http/http.dart' as http;
+import 'package:http_client/http_client.dart' as http_client;
+import 'package:http_client/pkg_http_adapter.dart' as pkg_http_adapter;
+import 'package:meta/meta.dart';
 
 import 'transport.dart';
 
-import 'dart:convert' as convert;
-
+/// The default transport implementation over HTTP.
 class HttpTransport implements Transport {
   final Uri _uri;
   final http.Client _httpClient;
   final Duration _timeout;
   final BasicAuth _basicAuth;
+  final bool _shouldCloseClient;
 
-  HttpTransport(this._httpClient, this._uri,
-      {Duration timeout = const Duration(minutes: 1), BasicAuth basicAuth})
-      : _timeout = timeout,
+  /// Creates a new HttpTransport instance.
+  HttpTransport({
+    @required /* String | Uri */ dynamic url,
+    dynamic client,
+    Duration timeout = const Duration(minutes: 1),
+    BasicAuth basicAuth,
+  })  : _httpClient = _castClient(client),
+        _shouldCloseClient = client != null,
+        _uri = _castUri(url),
+        _timeout = timeout,
         _basicAuth = basicAuth;
 
   @override
@@ -25,38 +35,52 @@ class HttpTransport implements Transport {
     ];
     final newUri = _uri.replace(
         pathSegments: pathSegments, queryParameters: request.params);
-    final rs = await _httpClient
-        .send(http.Request(
-          request.method,
-          newUri,
-          headers: _mergeHeader(request.headers),
-          body: request.bodyText,
-        ))
-        .timeout(_timeout);
-    return Response(rs.statusCode, await rs.readAsString());
+    final rq = http.Request(request.method, newUri);
+    if (_basicAuth != null) {
+      rq.headers.addAll(_basicAuth.toMap());
+    }
+    if (request.headers != null) {
+      rq.headers.addAll(request.headers);
+    }
+    if (request.bodyText != null) {
+      rq.body = request.bodyText;
+    }
+    final rs = await _httpClient.send(rq).timeout(_timeout);
+    return Response(rs.statusCode, await rs.stream.bytesToString());
   }
 
   @override
   Future close() async {
-    await _httpClient.close();
-  }
-
-  Map<String, String> _mergeHeader(Map<String, String> headerToMerge) {
-    if (_basicAuth != null) {
-      final headers = Map<String, String>.from(_basicAuth.toMap());
-      headers.addAll(headerToMerge);
-      return headers;
+    if (_shouldCloseClient) {
+      _httpClient.close();
     }
-    return headerToMerge;
   }
 }
 
 class BasicAuth {
   final String username;
   final String password;
+
   BasicAuth(this.username, this.password);
+
   Map<String, String> toMap() {
     final up = convert.utf8.encode('$username:$password');
     return {'Authorization': 'Basic ${convert.base64Encode(up)}'};
   }
+}
+
+Uri _castUri(dynamic url) {
+  ArgumentError.checkNotNull(url, 'url');
+  if (url is String) return Uri.parse(url);
+  if (url is Uri) return url;
+  throw ArgumentError('Unknown URI type: $url');
+}
+
+http.Client _castClient(dynamic client) {
+  if (client == null) return http.Client();
+  if (client is http.Client) return client;
+  if (client is http_client.Client) {
+    return pkg_http_adapter.PkgHttpAdapter(client);
+  }
+  throw ArgumentError('Unknown HTTP client: $client.');
 }