+++ lib/src/transport.dart
@@ -1,4 +1,3 @@
-import 'dart:async';
 import 'dart:convert' as convert;
 
 import 'package:meta/meta.dart';